Added big int library

+++ src/bailliePsw/bailliePsw.hpp
@@ -2,14 +2,15 @@
 #pragma once
 
 #include "millerRabin/millerRabin.hpp"
-#include "primalityTest/primalityTest.hpp"
+// #include "primalityTest/primalityTest.hpp"
+// #include "helpers.hpp"
 
 class BailliePSW : public PrimalityTest {
 public:
   BailliePSW() = default;
   ~BailliePSW() = default;
 
-  bool isPrime(__int128 n) override;
+  bool isPrime(BigInt n) override;
 
 private:
   MillerRabin millerRabin{1};
